Move bcf/errors into crate-wide errors
