+++ src/bcf/buffer.rs
@@ -6,7 +6,7 @@
 use std::collections::{vec_deque, VecDeque};
 use std::mem;
 
-use crate::bcf::errors::Result;
+use crate::errors::Result;
 use crate::bcf::{self, Read};
 
 /// A buffer for BCF records. This allows access regions in a sorted BCF file while iterating
