+++ src/errors.rs
@@ -7,13 +7,11 @@ pub type Result<T, E = Error> = std::result::Result<T, E>;
 
 #[derive(Error, Debug, PartialEq)]
 pub enum Error {
+		// General errors
     #[error("file not found: {path}")]
     FileNotFound { path: PathBuf },
     #[error("invalid (non-unicode) characters in path")]
     NonUnicodePath,
-    #[error("The given position is too large to be converted to i64")]
-    PositionTooLarge,
-
     #[error("failed to fetch region")]
     Fetch,
     #[error("error seeking to offset")]
@@ -21,6 +19,11 @@ pub enum Error {
     #[error("sequence {sequence} not found in index")]
     UnknownSequence { sequence: String },
 
+		// Errors for faidx
+    #[error("The given position is too large to be converted to i64")]
+    PositionTooLarge,
+
+		// Errors for Tbx
     #[error("previous iterator generation failed")]
     NoIter,
     #[error("truncated tabix record")]
@@ -30,6 +33,7 @@ pub enum Error {
     #[error("error setting threads for for file reading")]
     SetThreads,
 
+		// Errors for BAM
     #[error("error parsing CIGAR string: {msg}")]
     ParseCigar { msg: String },
     #[error("unexpected CIGAR operation: {msg}")]
@@ -50,7 +54,6 @@ pub enum Error {
     BamInvalidRecord,
     #[error("truncated record in SAM/BAM/CRAM file")]
     BamTruncatedRecord,
-
     #[error("format of SAM files are not indexable")]
     NotIndexable,
     #[error("failed to write BAM/CRAM index (out of disk space?)")]
@@ -59,4 +62,37 @@ pub enum Error {
     BuildIndex,
     #[error("failed to create SAM/BAM/CRAM pileup")]
     Pileup,
+
+
+		// Errors for BCF
+    #[error("error allocating internal data structure for BCF/VCF reader (out of memory?)")]
+    AllocationError,
+    #[error("failed to open BCF/VCF from {target:?}")]
+    Open { target: String },
+    #[error("invalid record in BCF/VCF file")]
+    InvalidRecord,
+    #[error("tag {tag} undefined in BCF/VCF header")]
+    UndefinedTag { tag: String },
+    #[error("unexpected type for tag {tag} BCF/VCF file")]
+    UnexpectedType { tag: String },
+    #[error("tag {tag} missing from record {record} in BCF/VCF file")]
+    MissingTag { tag: String, record: String },
+    #[error("error setting tag {tag} in BCF/VCF record (out of memory?)")]
+    SetTag { tag: String },
+    #[error("ID {rid} not found in BCF/VCF header")]
+    UnknownRID { rid: u32 },
+    #[error("contig {contig} not found in BCF/VCF header")]
+    UnknownContig { contig: String },
+    #[error("ID {id} not found in BCF/VCF header")]
+    UnknownID { id: String },
+    #[error("sample {name} not found in BCF/VCF header")]
+    UnknownSample { name: String },
+    #[error("duplicate sample names given for subsetting BCF/VCF")]
+    DuplicateSampleNames,
+    #[error("failed to set values in BCF/VCF record (out of memory?)")]
+    SetValues,
+    #[error("failed to remove alleles in BCF/VCF record")]
+    RemoveAlleles,
+    #[error("error seeking to {contig:?}:{start} in indexed BCF/VCF file")]
+    BcfSeek { contig: String, start: u64 }
 }
